v5/requirements.txt
# SGAPS-MAE Requirements
# Server-Guided Adaptive Pixel Sampling MAE for Game Session Replay

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# Configuration and utilities
hydra-core>=1.3.2
omegaconf>=2.3.0
PyYAML>=6.0.0

# Image processing
pillow>=10.0.0
opencv-python>=4.8.0

# Visualization and logging
matplotlib>=3.7.0
tensorboard>=2.14.0
tqdm>=4.66.0

# Data handling
h5py>=3.9.0

# Compression
zlib-state>=0.1.0
